run autoresearch cases in live e2e harness
